Fix: convert numpy float64 to native Python float to prevent psycopg2 crash
Root cause: _recency_score uses np.exp() which returns np.float64. This
value propagates through the final score calculation and round() preserves
the numpy type. When psycopg2 formats this for PostgreSQL, it renders as
np.float64(value), which PostgreSQL interprets as schema np not existing:

  psycopg2.errors.InvalidSchemaName: schema np does not exist

This caused db.commit() to fail, rolling back the entire transaction
(including correctly classified stories), so the old wrong stories
remained in the database.

The fix converts score to a native Python float in rank_clusters, and
also fixes _recency_score to return float instead of np.float64, since
its type annotation says -> float.

+++ backend/services/rank_news.py
@@ -111,7 +111,7 @@ def _recency_score(latest: datetime) -> float:
         return 0.0
     # Exponential decay: score = e^(-0.1 * (age-2))
     # This gives: ~82% at 4h, ~67% at 6h, ~37% at 12h, ~14% at 24h
-    return max(0.0, np.exp(-0.1 * (age_hours - 2)))
+    return float(max(0.0, np.exp(-0.1 * (age_hours - 2))))
 
 
 
@@ -160,7 +160,7 @@ def rank_clusters(clusters: list[list[dict]]) -> list[dict]:
 
         scored.append({
             "cluster": cluster,
-            "score": round(final, 4),
+            "score": float(round(final, 4)),
             "article_count": size,
             "latest_at": latest.isoformat(),
             "sources": sources,
